rootio_web/requirements.txt
-i https://pypi.python.org/simple
alembic==0.6.0
anyjson==0.3.3
apscheduler==2.0.2
argparse==1.2.1
babel==1.3
backports.shutil-get-terminal-size==1.0.0; python_version == '2.7'
beautifulsoup==3.2.1
blinker==1.3
boto3==1.7.65
botocore==1.10.65
coaster==0.4.1
decorator==3.4.0
docflow==0.3.1
docutils==0.14
ecdsa==0.11
enum34==1.1.6; python_version == '2.7'
fabric==1.7.0
flask-admin==1.0.7
flask-assets==0.10
flask-babel==0.9
flask-cache==0.12
flask-gravatar==0.5.0
flask-login==0.2.6
flask-mail==0.9.0
flask-openid==1.2.4
flask-restless==0.12.1
flask-script==0.5.3
flask-sqlalchemy==1.0
flask-testing==0.6.1
flask-wtf==0.9.1
flask==0.10.1
futures==2.2.0
infinity==1.3
intervals==0.3.1
ipdb==0.11
ipython-genutils==0.2.0
ipython==5.7.0
isodate==0.5.0
itsdangerous==0.24
jinja2==2.7.1
jmespath==0.9.3
mako==0.8.1
markdown==2.5.2
markupsafe==0.23
ofs==0.4.2
packaging==17.1
pairtree==0.7.1-t
paramiko==1.15.2
pathlib2==2.3.2; python_version in '2.6 2.7 3.2 3.3'
pathlib==1.0.1
pexpect==4.6.0; sys_platform != 'win32'
pickleshare==0.7.4
plivohelper==0.3.5
pluggy==0.7.1
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.4.32
pycrypto==2.6.1
pygments==2.0.1
pyparsing==2.2.0
pytest==3.0.5
python-dateutil==2.2
python-openid==2.2.5
pytz==2014.10
pyzmq==14.2.0
requests==2.5.1
s3transfer==0.1.13
scandir==1.7; python_version < '3.5'
semantic-version==2.3.1
simplegeneric==0.8.1
simplejson==3.6.5
six==1.11.0
speaklater==1.3
sqlalchemy-utils==0.25.0
sqlalchemy==0.9.3
tox==3.1.2
traitlets==4.3.2
tzlocal==1.1.2
validators==0.7
virtualenv==16.0.0
waitress==1.0.1
wcwidth==0.1.7
webassets==0.10.1
werkzeug==0.9.3
wsgiref==0.1.2
wtforms-alchemy==0.7.12
wtforms-components==0.9.1
wtforms==1.0.4
yapsy==1.10.423
